Update common infrastructure to implement C++11 utility functions
