+++ resource/fastDDS/CMakeLists.txt
@@ -21,46 +21,7 @@ set(CMAKE_LIBRARY_OUTPUT_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/bin/Linux)
 set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/bin/Linux)
 set(CMAKE_ARCHIVE_OUTPUT_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/bin/Linux)
 
-# This can be removed once we stop depending on Micro Osapi to build.
-set(RTI_LIB_PREFIX rti_me)
-set(RTI_LIB_SUFFIX)
-
-IF(NOT (DEFINED RTI_SHARED_LIB))
-    IF (${CMAKE_BUILD_TYPE} MATCHES "[Rr]elease")
-        set(RTI_LIB_SUFFIX "z")
-        set(RTI_DEBUG_BUILD FALSE)
-        ADD_DEFINITIONS(-DNDEBUG)
-    ELSE()
-        set(RTI_LIB_SUFFIX "zd")
-        set(RTI_DEBUG_BUILD TRUE)
-    ENDIF()
-ELSE()
-    IF (${CMAKE_BUILD_TYPE} MATCHES "[Rr]elease")
-        set(RTI_LIB_SUFFIX "")
-        set(RTI_DEBUG_BUILD FALSE)
-        ADD_DEFINITIONS(-DNDEBUG)
-    ELSE()
-        set(RTI_LIB_SUFFIX "d")
-        set(RTI_DEBUG_BUILD TRUE)
-    ENDIF()
-ENDIF()
-
-set(MICRO_C_LIBS 
-    rti_me_cpp${RTI_LIB_SUFFIX}
-    rti_me_netioshmem${RTI_LIB_SUFFIX}
-    rti_me_netiosdm${RTI_LIB_SUFFIX}
-    rti_me_rhsm${RTI_LIB_SUFFIX}
-    rti_me_whsm${RTI_LIB_SUFFIX}
-    rti_me_discdpse${RTI_LIB_SUFFIX}
-    rti_me_discdpde${RTI_LIB_SUFFIX}
-    rti_me${RTI_LIB_SUFFIX})
-
 set(FASTDDS_LIBS fastcdr fastrtps)
-
-# This will not be needed once we take out the Micro dependency
-set(PLATFORM_LIBS dl;m;pthread;)
-INCLUDE_DIRECTORIES($ENV{RTIMEHOME}/include $ENV{RTIMEHOME}/include/rti_me)
-LINK_DIRECTORIES(/performance/rtimehome/rti_me.2.0/lib/x64Linux4gcc7.3.0)
 
 ADD_EXECUTABLE(FastDDSperftest_cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/srcCpp/fastDDS/perftest.cxx
@@ -77,4 +38,4 @@ ADD_EXECUTABLE(FastDDSperftest_cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/srcCpp/FileDataLoader.cxx
         ${CMAKE_CURRENT_SOURCE_DIR}/srcCpp/perftest_publisher.cxx)
 
-TARGET_LINK_LIBRARIES(FastDDSperftest_cpp  ${MICRO_C_LIBS} ${FASTDDS_LIBS} ${PLATFORM_LIBS})
+TARGET_LINK_LIBRARIES(FastDDSperftest_cpp ${FASTDDS_LIBS})
