+++ srcCpp/fastDDS/Infrastructure_fastDDS.cxx
@@ -11,127 +11,6 @@ using namespace eprosima::fastdds::rtps;
 using namespace eprosima::fastdds::dds;
 
 /********************************************************************/
-/* Perftest Clock class */
-
-PerftestClock::PerftestClock()
-{
-
-}
-
-PerftestClock::~PerftestClock()
-{
-}
-
-PerftestClock &PerftestClock::getInstance()
-{
-    static PerftestClock instance;
-    return instance;
-}
-
-unsigned long long PerftestClock::getTime()
-{
-    clock_gettime(CLOCK_MONOTONIC, &timeStruct);
-    return (timeStruct.tv_sec * ONE_MILLION) + timeStruct.tv_nsec/1000;
-}
-
-void PerftestClock::milliSleep(unsigned int millisec)
-{
-    OSAPI_Thread_sleep(millisec);
-}
-
-void PerftestClock::sleep(const struct DDS_Duration_t& sleep_period)
-{
-    NDDSUtility::sleep(sleep_period);
-}
-
-/********************************************************************/
-/* Get Connext Micro functions */
-
-const std::string GetDDSVersionString()
-{
-    return "Eprosima FastDDS";
-}
-
-void PerftestConfigureVerbosity(int verbosityLevel)
-{
-
-    OSAPI_LogVerbosity_T verbosity = OSAPI_LOG_VERBOSITY_ERROR;
-    switch (verbosityLevel) {
-        case 0: fprintf(stderr, "[Error]: Cannot set verbosity to SILENT\n");
-                break;
-        case 1: Log::SetVerbosity(Log::Error);
-                fprintf(stderr, "Setting verbosity to ERROR\n");
-                break;
-        case 2: Log::SetVerbosity(Log::Warning);
-                fprintf(stderr, "Setting verbosity to WARNING\n");
-                break;
-        case 3: Log::SetVerbosity(Log::Info);
-                fprintf(stderr, "Setting verbosity to INFO\n");
-                break;
-        default: fprintf(stderr,
-                    "[Error]: Invalid value for the verbosity parameter. Using default\n");
-                break;
-    }
-}
-
-/********************************************************************/
-/* Micro OnSpawed Method */
-
-struct PerftestMicroThreadOnSpawnedMethod
-{
-    MicroThreadOnSpawnedMethod method;
-    void *thread_param;
-
-};
-
-static int perftestMicroThreadRoutine(struct OSAPI_ThreadInfo *thread_info)
-{
-    PerftestMicroThreadOnSpawnedMethod *data = (PerftestMicroThreadOnSpawnedMethod *)thread_info->user_data;
-
-    data->method(data->thread_param);
-
-    delete(data);
-
-    thread_info->stop_thread = RTI_TRUE;
-    return 1;
-}
-
-/********************************************************************/
-/* Thread Related functions */
-
-struct PerftestThread* PerftestThread_new(
-        const char *name,
-        int threadPriority,
-        int threadOptions,
-        MicroThreadOnSpawnedMethod method,
-        void *threadParam)
-{
-    struct OSAPI_ThreadProperty prio = OSAPI_ThreadProperty_INITIALIZER;
-    PerftestMicroThreadOnSpawnedMethod *data = new PerftestMicroThreadOnSpawnedMethod();
-    data->method = method;
-    data->thread_param = threadParam;
-
-    struct OSAPI_Thread *thread = NULL;
-    thread = OSAPI_Thread_create(
-                name,
-                &prio,
-                perftestMicroThreadRoutine,
-                data,
-                NULL);
-    if (!OSAPI_Thread_start(thread)) {
-        return NULL;
-    }
-    return thread;
-}
-
-void PerftestThread_delete(struct PerftestThread* thread)
-{
-    if (!OSAPI_Thread_destroy(thread)) {
-        printf("Error deleting thread");
-    }
-}
-
-/********************************************************************/
 /* Transport related functions */
 
 bool configure_udpv4_transport(
